Restructured all files creating their specific folders.
